requirements.txt
playwright==1.58.0
pytest>=8.0
python-dotenv>=1.0
